requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.6
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
flight-optimizer==0.2.0
gunicorn==20.0.4
haversine==2.2.0
idna==2.10
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-mock==3.1.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.1.0
